factor out extensions into own classes
