+++ main/src/main/java/cgeo/geocaching/utils/gpx/GPXParser.java
@@ -46,18 +46,6 @@ public class GPXParser {
 
     private IGPXParseHooks hooks;
 
-    /** Controls how wptTypes are parsed. */
-    public enum ParseMode {
-        /** Parse/collect complete information. */
-        FULL,
-        /** Parse only basic data per wptType (name + coordinate) */
-        COORDINATES_ONLY,
-        /** Do not collect wptType data.*/
-        SKIP,
-        /** Flag to use to abort parsing completely */
-        ABORT
-    }
-
     private static class AbortException extends RuntimeException {
 
     }
