+++ main/src/main/java/cgeo/geocaching/utils/gpx/IGPXExtension.java
@@ -1,0 +1,23 @@
+package cgeo.geocaching.utils.gpx;
+
+import cgeo.geocaching.log.LogEntry;
+import cgeo.geocaching.models.Geocache;
+import cgeo.geocaching.models.Waypoint;
+import cgeo.geocaching.utils.xml.XmlNode;
+
+import java.util.List;
+
+public interface IGPXExtension {
+
+    default void enrichGeocache(XmlNode wptNode, Geocache cache) {
+        //do nothing
+    }
+
+    default void enrichWaypoint(XmlNode wptNode, Waypoint waypoint) {
+        //do nothing
+    }
+
+    default List<LogEntry> extractLogs(XmlNode wptNode, Geocache cache) {
+        return null;
+    }
+}
